+++ dflib/src/main/java/org/dflib/slice/ConditionalRowSet.java
@@ -2,6 +2,7 @@ package org.dflib.slice;
 
 import org.dflib.BooleanSeries;
 import org.dflib.DataFrame;
+import org.dflib.Exp;
 import org.dflib.IntSeries;
 import org.dflib.RowSet;
 import org.dflib.Series;
@@ -15,34 +16,32 @@ public class ConditionalRowSet extends BaseRowSet {
     private final BooleanSeries conditionalIndex;
 
     public ConditionalRowSet(DataFrame source, BooleanSeries conditionalIndex) {
-        this(source, -1, null, null, conditionalIndex);
+        this(source, null, null, null, conditionalIndex);
     }
 
     protected ConditionalRowSet(
             DataFrame source,
-            int expansionColumn,
+            Exp<?> expansionExp,
             int[] uniqueColumns,
             Sorter[] sorters,
             BooleanSeries conditionalIndex) {
-        super(source, expansionColumn, uniqueColumns, sorters);
+        super(source, expansionExp, uniqueColumns, sorters);
         this.conditionalIndex = conditionalIndex;
     }
 
     @Override
-    public RowSet expand(int columnPos) {
-        return this.expansionColumn != columnPos
-                ? new ConditionalRowSet(source, columnPos, uniqueKeyColumns, sorters, conditionalIndex)
-                : this;
+    public RowSet expand(Exp<?> splitExp) {
+        return new ConditionalRowSet(source, splitExp, uniqueKeyColumns, sorters, conditionalIndex);
     }
 
     @Override
     public RowSet unique(int... uniqueKeyColumns) {
-        return new ConditionalRowSet(source, expansionColumn, uniqueKeyColumns, sorters, conditionalIndex);
+        return new ConditionalRowSet(source, expansionExp, uniqueKeyColumns, sorters, conditionalIndex);
     }
 
     @Override
     public RowSet sort(Sorter... sorters) {
-        return new ConditionalRowSet(source, expansionColumn, uniqueKeyColumns, sorters, conditionalIndex);
+        return new ConditionalRowSet(source, expansionExp, uniqueKeyColumns, sorters, conditionalIndex);
     }
 
     @Override