+++ dflib/src/test/java/org/dflib/RowSet_Expand_Merge_ExpTest.java
@@ -44,6 +44,32 @@ public class RowSet_Expand_Merge_ExpTest {
     }
 
     @Test
+    public void byIndex_ExpQL_NewColumn() {
+        DataFrame df = DataFrame.foldByRow("a", "b")
+                .of(
+                        1, "x1,x2", // <--
+                        2, "y1,y2",
+                        3, "m1,m2") // <--
+                .rows(Series.ofInt(0, 2))
+                .expand("split(str(b), ',') as d")
+
+                // the expansion added a column, so the number of expressions must match the expanded width
+                .merge(
+                        $int("a").add(1),
+                        $str("b"),
+                        $str("d").mapVal(String::toUpperCase)
+                );
+
+        new DataFrameAsserts(df, "a", "b", "d")
+                .expectHeight(5)
+                .expectRow(0, 2, "x1,x2", "X1")
+                .expectRow(1, 2, "x1,x2", "X2")
+                .expectRow(2, 2, "y1,y2", null)
+                .expectRow(3, 4, "m1,m2", "M1")
+                .expectRow(4, 4, "m1,m2", "M2");
+    }
+
+    @Test
     public void byCondition() {
         DataFrame df = DataFrame.foldByRow("a", "b", "c")
                 .of(